Added script for locating a relative forge-enabled project.
This is done by ascending the file system and checking for a .forge folder. If the /Users/Abe.Walters directory is reached, the task terminates in an error with no project found, otherwise it will execute on the first project found.

In addition, FORGE_TMP was changed to the more accurate FORGE_CACHE.